Fix `cabal test` not solving for test suite in CI
For some reason, it's not required for the solver to include the test
suite when running `cabal test`. This means that there is a chance that
`cabal test` can reach a local maxima of dependencies that bizzarely
doesn't contain the test suite itself. This is tracked in:

    https://github.com/haskell/cabal/issues/7883

The solution is either to add tests: True to cabal.project, or call
`cabal test` with the --enable-tests flag. This commit does the latter
in the CI, but eventually it'd be nice to get rid of the confusing flag.