Add opus_compare to repo, call it from Haskel during tests
The goal is to have opus_compare in source code format (for minimal
footprint), and ideally in unaltered form from the Xiph repo for easy
updates/reproducibility.

We (ab)use the fact that Cabal can compile normal C code. Specifically,
we build opus_compare.c and link to it from Haskell and call into its
main function using an FFI wrapper (OpusCompare.hs). To avoid having
multiple definitions of main(), we use the cc-options field in the
cabal file to rename the main function in opus_compare.c to
opus_compare_main.

This is a bit of a hack, but it works and is simple. No more need to run
gcc separately and have a separate executable, or to need to install
opus-tools locally!